requirement.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-rest-swagger==2.1.0
djangorestframework==3.5.3
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mysqlclient==1.3.14
numpy==1.16.0
openapi-codec==1.3.2
pandas==0.23.4
PyMySQL==0.9.3
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
requests==2.21.0
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1